Implemented choppy GetY function

+++ Terrain.h
@@ -38,7 +38,7 @@ public:
 
 	void Render(ID3D11DeviceContext* deviceContext);
 
-	int GetY(int x, int z);
+	float GetY(int x, int z);
 
 	int GetIndexCount();
 };
